create file for select choises and change message to display for admin #183
