+++ client/src/components/application/NonlinearStepper/questions.js
@@ -1,0 +1,64 @@
+const GazaAreas = [
+  'North of Gaza Strip',
+  'Gaza City',
+  'Middle Area of Gaza',
+  'Khan Younis',
+  'Rafah',
+];
+const GazaUniversities = [
+  "I didn't go to university",
+  'Islamic University of Gaza',
+  'Al Azhar University',
+  'Palestine University',
+  'UCAS',
+  'Al Quds Open University',
+  'University College of Science and Technology',
+  'Al Aqsa University',
+  'Gaza University',
+  'I went to university outside of Gaza',
+];
+
+const KhalilUniversities = [
+  "I didn't go to university",
+  'Palestine Polytechnic University (PPU)',
+  'Hebron University',
+  'Bethlehem University',
+  'Birzeit University',
+  'Al Najah University',
+  'Arab American University',
+  'Al Quds Open University',
+  'Al Quds University',
+  'Palestine Technical University - Kadoori',
+  'IT College - Vocational Center',
+  'University outside of Palestine',
+  'Other',
+];
+const advertisement = [
+  'Gaza Sky Geeks Social Media (Facebook, Twitter, Instagram)',
+  'Friend or family member who is a Code Academy graduate',
+  'I came to a GSG workshop or event (such as “Coding for Everyone”)',
+  'Online / browsing the web',
+  'Jobs.ps',
+  'Other',
+];
+const codeExperience = [
+  'No Experience',
+  'Some experience as hobby',
+  'A lot of experience as a hobby',
+  'I have a university degree in Computer Science or another related field',
+  'I have had a paid job as a developer.',
+];
+const primaryMotivationForm = [
+  'I want to gain employment/earn income as a developer.',
+  "I want to learn something new - I don't know if I want a coding career",
+  'I want to be part of the GSG and Code Academy community',
+];
+
+export default {
+  GazaAreas,
+  GazaUniversities,
+  KhalilUniversities,
+  advertisement,
+  codeExperience,
+  primaryMotivationForm,
+};
