+++ client/src/utils/application/nLinearStepperValidation.js
@@ -10,18 +10,24 @@ const generalInfosSchema = yup.object().shape({
   age: yup.string().required('Age is required').trim(),
   address: yup.string().required('Address is required'),
   motivation: yup.string().required(`Insert motivation`).trim(),
-  specialization: yup.string().required(`specialization`).trim(),
-  university: yup.string().required(`university`).trim(),
-  codingExperience: yup.string().required(`codingExperience`).trim(),
-  primaryMotivation: yup.string().required(`primaryMotivation`).trim(),
-  caReading: yup.string().required(`caReading`).trim(),
+  specialization: yup.string().required(`Insert your Specialization`).trim(),
+  university: yup.string().required(`Select your university`).trim(),
+  codingExperience: yup
+    .string()
+    .required(`Select your coding experience`)
+    .trim(),
+  primaryMotivation: yup
+    .string()
+    .required(`Select your primary motivation first`)
+    .trim(),
+  caReading: yup.string().required(`Select how you here about Us first`).trim(),
   cvLink: yup
     .string()
     .matches(
       /^((https?):\/\/)?(www.)?[a-z0-9]+(\.[a-z]{2,}){1,3}(#?\/?[a-zA-Z0-9#]+)*\/?(\?[a-zA-Z0-9-_]+=[a-zA-Z0-9-%]+&?)?$/,
       'Enter correct url!'
     )
-    .required(`cvLink`)
+    .required(`Insert Your Cv Link first`)
     .trim(),
 });
